requirements.txt
numpy==2.3.2
pandas==2.3.2
plotly==6.2.0
streamlit==1.49.1
scikit-learn==1.7.0
google-generativeai
joblib
matplotlib
seaborn